requirements.txt
tensorflow
pandas
matplotlib
kafka-python